+++ common/src/main/java/mb/common/util/CollectionView.java
@@ -7,6 +7,8 @@ import java.util.ArrayList;
 import java.util.Collection;
 import java.util.Collections;
 import java.util.Objects;
+import java.util.stream.Collectors;
+import java.util.stream.Stream;
 
 /**
  * A wrapper around a {@link Collection collection} with read-only operations. Only {@link Serializable serializable}
@@ -40,6 +42,11 @@ public class CollectionView<E> extends BaseCollectionView<E, Collection<? extend
         return new CollectionView<>(collection);
     }
 
+    public static <E> CollectionView<E> of(Stream<? extends E> elements) {
+        final ArrayList<E> list = elements.collect(Collectors.toCollection(ArrayList::new));
+        return new CollectionView<>(list);
+    }
+
 
     public static <E> CollectionView<E> copyOf(Iterable<? extends E> elements) {
         final ArrayList<E> list = new ArrayList<>();
@@ -49,6 +56,10 @@ public class CollectionView<E> extends BaseCollectionView<E, Collection<? extend
 
     public static <E> CollectionView<E> copyOf(Collection<? extends E> collection) {
         return new CollectionView<>(new ArrayList<>(collection));
+    }
+
+    public static <E> CollectionView<E> copyOf(CollectionView<? extends E> collection) {
+        return new CollectionView<>(new ArrayList<>(collection.collection));
     }