+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project are documented in this file, based on [Keep 
 ## [Unreleased]
 ### Added
 - `MultiMap.valueStream`
+- `CollectionView.of(Stream)`
+- `CollectionView.copyOf(CollectionView)`
+- `ListView.of(Stream)`
+
 
 ## [0.10.1] - 2021-11-23
 ### Changed